multiav_prj/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-debug-toolbar==4.2.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
httpie==3.2.2
idna==3.4
inflection==0.5.1
Markdown==3.4.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pymemcache==4.0.0
PySocks==1.7.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.3
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.5
